refactor(ui): split monolith files into focused modules
InsightsDrawer (1175→270 lines):
- Extract 9 tab components to components/document/tabs/
- TabButton, IndexTab, StatsTab, ChunkSummaryTab, CoherenceTab,
  AuditTab, IssueList, NotesTab, GlossaryTab

DocumentView (804→548 lines):
- Extract StageTraceDialog, PaneSearch, InlineStatusBadge
- Extract useDocumentViewState hook to hooks/

usePipeline (901→593 lines):
- Extract pipeline/blobContext.ts (pure utils)
- Extract pipeline/runJudge.ts (judge logic)
- Extract usePipelineAudit.ts (audit hooks)

docs: update ARCHITECTURE.md store table
