+++ src/components/document/InsightsDrawer.tsx
@@ -1,59 +1,35 @@
 import {
-  AlertCircle,
-  AlertTriangle,
   BarChart2,
-  BookOpen,
   BookText,
-  CheckCheck,
-  CheckCircle2,
-  Circle,
-  Clock,
-  Crosshair,
-  Cpu,
-  FileText,
-  FlaskConical,
-  Gauge,
+  Brain,
   Link2,
   List,
-  Loader2,
-  MessageCircle,
   NotebookText,
   PanelRight,
-  RefreshCcw,
-  ScanLine,
   Search,
-  Brain,
   ShieldCheck,
   TerminalSquare,
   X,
 } from 'lucide-react';
-import { useVirtualizer } from '@tanstack/react-virtual';
 import { motion, AnimatePresence } from 'motion/react';
 import { useTranslation } from 'react-i18next';
-import { type KeyboardEvent, useEffect, useMemo, useRef, useState } from 'react';
+import { type KeyboardEvent, useEffect, useRef } from 'react';
 import { useUiStore, type InsightsDrawerTab, type ChunkDrawerTab } from '../../stores/uiStore';
 import { useChunksStore } from '../../stores/chunksStore';
-import { usePhraseMemoryStore } from '../../stores/phraseMemoryStore';
 import { MemoryTab } from './MemoryTab';
 import { usePipelineStore } from '../../stores/pipelineStore';
-import { usePricingStore } from '../../stores/pricingStore';
-import { indexPad, qualityLabelKey, qualityTone, calculateCompositeQuality } from '../../utils';
-import {
-  formatCacheHitRate,
-  formatDurationMs,
-  formatUsd,
-  summarizeChunkUsage,
-  summarizeGlobalUsage,
-  type OperationLogRunSummary,
-  type ScopeBreakdownEntry,
-} from '../../utils/operationLogStats';
-import { useOperationLogStore } from '../../stores/operationLogStore';
 import { useChunkWatchdog } from '../../hooks/useChunkWatchdog';
-import { usePipeline } from '../../hooks/usePipeline';
 import { OperationsTab } from './OperationsTab';
 import { SearchTab } from './SearchTab';
-import type { TranslationChunk } from '../../types';
 import { IconButton, Tooltip } from '../ui';
+import { TabButton } from './tabs/TabButton';
+import { IndexTab } from './tabs/IndexTab';
+import { StatsTab } from './tabs/StatsTab';
+import { ChunkSummaryTab } from './tabs/ChunkSummaryTab';
+import { CoherenceTab } from './tabs/CoherenceTab';
+import { AuditTab } from './tabs/AuditTab';
+import { NotesTab } from './tabs/NotesTab';
+import { GlossaryTab } from './tabs/GlossaryTab';
 
 interface InsightsDrawerProps {
   onReauditChunk: (chunkId: string) => void;
@@ -97,12 +73,6 @@ const CHUNK_TAB_PANEL_IDS: Record<ChunkDrawerTab, string> = {
   memory: 'chunk-tab-panel-memory',
 };
 
-const QUALITY_TONE_COLOR: Record<ReturnType<typeof qualityTone>, string> = {
-  strong: 'text-editorial-success',
-  ok: 'text-editorial-warning',
-  weak: 'text-editorial-accent',
-};
-
 export function InsightsDrawer({ onReauditChunk, onRunCoherenceAudit }: InsightsDrawerProps) {
   const { t } = useTranslation();
   const showDocumentDrawer = useUiStore((state) => state.showDocumentDrawer);
@@ -125,7 +95,6 @@ export function InsightsDrawer({ onReauditChunk, onRunCoherenceAudit }: Insights
   const currentChunkIndex = currentChunk ? chunks.findIndex((c) => c.id === currentChunk.id) : -1;
 
   const { stuckChunkIds, cancelStuckChunk } = useChunkWatchdog();
-  const { rerunChunkWithMemory } = usePipeline();
   const { config } = usePipelineStore();
   const hasGlossary = !!config.assignedGlossaryId && config.glossary.length > 0;
 
@@ -454,722 +423,6 @@ export function InsightsDrawer({ onReauditChunk, onRunCoherenceAudit }: Insights
         )}
       </AnimatePresence>
 
-    </div>
-  );
-}
-
-// ── Tab Button ─────────────────────────────────────────────────────────────
-
-interface TabButtonProps {
-  buttonId: string;
-  active: boolean;
-  disabled?: boolean;
-  onClick: () => void;
-  onKeyDown: (event: KeyboardEvent<HTMLButtonElement>) => void;
-  label: string;
-  icon: React.ReactNode;
-  controls: string;
-  buttonRef: (element: HTMLButtonElement | null) => void;
-}
-
-function TabButton({ buttonId, active, disabled, onClick, onKeyDown, label, icon, controls, buttonRef }: TabButtonProps) {
-  return (
-    <IconButton
-      id={buttonId}
-      ref={buttonRef}
-      size="lg"
-      tone={active ? 'accent' : 'default'}
-      role="tab"
-      aria-selected={active}
-      aria-controls={controls}
-      aria-disabled={disabled}
-      disabled={disabled}
-      tabIndex={active ? 0 : -1}
-      onClick={onClick}
-      onKeyDown={onKeyDown}
-      title={label}
-      ariaLabel={label}
-      tooltipSide="bottom"
-    >
-      {icon}
-    </IconButton>
-  );
-}
-
-// ── Index Tab ──────────────────────────────────────────────────────────────
-
-interface IndexTabProps {
-  panelId: string;
-  labelledBy: string;
-  chunks: TranslationChunk[];
-  currentChunkId: string | null;
-  isProcessing: boolean;
-  stuckChunkIds: Set<string>;
-  onSelect: (id: string) => void;
-  onCancelStuck: (chunkId: string) => void;
-}
-
-function IndexTab({ panelId, labelledBy, chunks, currentChunkId, isProcessing, stuckChunkIds, onSelect, onCancelStuck }: IndexTabProps) {
-  const { t } = useTranslation();
-  const matchesByChunk = usePhraseMemoryStore((s) => s.matchesByChunk);
-  const scrollRef = useRef<HTMLDivElement>(null);
-
-  const virtualizer = useVirtualizer({
-    count: chunks.length,
-    getScrollElement: () => scrollRef.current,
-    estimateSize: () => 80,
-    overscan: 5,
-  });
-
-  if (chunks.length === 0) {
-    return (
-      <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="flex flex-col items-center justify-center gap-3 px-6 py-12 text-center">
-        <List size={28} className="text-editorial-border" />
-        <p className="text-sm font-medium text-editorial-muted">{t('document.indexEmptyTitle')}</p>
-        <p className="text-xs leading-relaxed text-editorial-muted/70">{t('document.indexEmptyBody')}</p>
-      </div>
-    );
-  }
-
-  return (
-    <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} ref={scrollRef} className="flex-1 min-h-0 overflow-y-auto custom-scrollbar px-4 py-4">
-      <ul style={{ height: virtualizer.getTotalSize(), position: 'relative' }} className="w-full">
-        {virtualizer.getVirtualItems().map((virtualRow) => {
-          const chunk = chunks[virtualRow.index];
-          const index = virtualRow.index;
-          const isActive = chunk.id === currentChunkId;
-          const tone = qualityTone(chunk.judgeResult.status === 'completed' ? chunk.judgeResult.rating : null);
-          const wordCount = chunk.originalText.trim() ? chunk.originalText.trim().split(/\s+/).filter(Boolean).length : 0;
-          const isStuck = stuckChunkIds.has(chunk.id);
-
-          let statusIcon: React.ReactNode;
-          if (chunk.status === 'processing') {
-            statusIcon = isStuck
-              ? <Clock size={13} className="text-editorial-accent shrink-0" />
-              : <Loader2 size={13} className="animate-spin text-editorial-warning shrink-0" />;
-          } else if (chunk.status === 'completed') {
-            statusIcon = <CheckCircle2 size={13} className="text-editorial-success shrink-0" />;
-          } else if (chunk.status === 'preview') {
-            statusIcon = <FlaskConical size={13} className="text-editorial-muted shrink-0" />;
-          } else if (chunk.status === 'error') {
-            statusIcon = <AlertCircle size={13} className="text-editorial-accent shrink-0" />;
-          } else {
-            statusIcon = <Circle size={13} className="text-editorial-border shrink-0" />;
-          }
-
-          return (
-            <li
-              key={chunk.id}
-              data-index={virtualRow.index}
-              ref={virtualizer.measureElement}
-              style={{ position: 'absolute', top: virtualRow.start, left: 0, right: 0 }}
-              className="pb-2"
-            >
-              <div className={`rounded-2xl border transition-colors ${isActive ? 'border-editorial-charcoal bg-editorial-charcoal' : 'border-editorial-border bg-editorial-bg hover:border-editorial-charcoal/40'}`}>
-                <button type="button" onClick={() => onSelect(chunk.id)} className="w-full px-4 pt-3 pb-2 text-left">
-                  <div className="flex items-center gap-2">
-                    {statusIcon}
-                    <span className={`font-display text-sm italic ${isActive ? 'text-white' : 'text-editorial-accent'}`}>
-                      {indexPad(index + 1)}
-                    </span>
-                    <span className={`flex-1 line-clamp-2 text-[11px] leading-snug ${isActive ? 'text-white/80' : 'text-editorial-muted'}`}>
-                      {chunk.originalText.replace(/\s+/g, ' ').trim()}
-                    </span>
-                    <span className={`shrink-0 text-[10px] font-mono ${isActive ? 'text-white/50' : 'text-editorial-muted/60'}`}>
-                      {wordCount}w
-                    </span>
-                  </div>
-                  {chunk.judgeResult.status === 'completed' && (
-                    <div className={`mt-1.5 flex items-center gap-1.5 text-[10px] font-bold uppercase tracking-[0.2em] ${isActive ? 'text-white/70' : QUALITY_TONE_COLOR[tone]}`}>
-                      <span className={`inline-block h-1.5 w-1.5 rounded-full ${tone === 'strong' ? 'bg-editorial-success' : tone === 'ok' ? 'bg-editorial-warning' : 'bg-editorial-accent'}`} />
-                      {t(qualityLabelKey(chunk.judgeResult.rating))}
-                    </div>
-                  )}
-                  {chunk.translationLocked && (
-                    <div className={`mt-1.5 flex items-center gap-1 text-[10px] font-bold uppercase tracking-[0.18em] ${isActive ? 'text-white/80' : 'text-editorial-success'}`}>
-                      <CheckCheck size={12} />
-                      {t('document.translationLockedBadge')}
-                    </div>
-                  )}
-                  {(() => {
-                    const matchCount = matchesByChunk.get(chunk.id)?.matches.length ?? 0;
-                    if (matchCount === 0) return null;
-                    return (
-                      <div className={`mt-1.5 flex items-center gap-1 text-[10px] font-bold uppercase tracking-[0.18em] ${isActive ? 'text-white/80' : 'text-editorial-accent'}`}>
-                        <Brain size={11} />
-                        {t('memory.matchBadge', { count: matchCount })}
-                      </div>
-                    );
-                  })()}
-                </button>
-
-                {isStuck && chunk.status === 'processing' && (
-                  <div className={`flex items-center justify-between gap-2 border-t px-3 py-2 ${isActive ? 'border-white/10' : 'border-editorial-border/60'}`}>
-                    <div className={`flex items-center gap-1.5 text-[10px] font-bold uppercase tracking-[0.18em] ${isActive ? 'text-orange-200' : 'text-editorial-accent'}`}>
-                      <Clock size={11} />
-                      {t('document.watchdogStuck')}
-                    </div>
-                    <button
-                      type="button"
-                      onClick={(e) => { e.stopPropagation(); onCancelStuck(chunk.id); }}
-                      aria-label={t('document.watchdogCancel')}
-                      className={`rounded-full border px-2 py-0.5 text-[10px] font-medium transition-colors focus:outline-none focus-visible:ring-2 focus-visible:ring-editorial-accent ${isActive ? 'border-orange-300/40 text-orange-200 hover:bg-white/10' : 'border-editorial-accent/40 text-editorial-accent hover:bg-editorial-accent/10'}`}
-                    >
-                      {t('document.watchdogCancel')}
-                    </button>
-                  </div>
-                )}
-              </div>
-            </li>
-          );
-        })}
-      </ul>
-    </div>
-  );
-}
-
-// ── Stats Tab ──────────────────────────────────────────────────────────────
-
-interface StatsTabProps {
-  panelId: string;
-  labelledBy: string;
-  chunks: TranslationChunk[];
-}
-
-function StatsTab({ panelId, labelledBy, chunks }: StatsTabProps) {
-  const { t } = useTranslation();
-  const pricingOverrides = usePricingStore((s) => s.overrides);
-  const logEntries = useOperationLogStore((s) => s.entries);
-  const usageSummary = useMemo(
-    () => summarizeGlobalUsage(logEntries, pricingOverrides),
-    [logEntries, pricingOverrides],
-  );
-
-  const sourceWords = chunks.reduce((acc, c) => acc + countWords(c.originalText), 0);
-  const translatedWords = chunks.reduce((acc, c) => acc + countWords(c.currentDraft || ''), 0);
-  const coverageRatio = sourceWords > 0 ? Math.round((translatedWords / sourceWords) * 100) : 0;
-  const total = chunks.length;
-  const idleCount = chunks.filter((c) => c.status === 'ready').length;
-  const processingCount = chunks.filter((c) => c.status === 'processing').length;
-  const completedCount = chunks.filter((c) => c.status === 'completed').length;
-  const previewCount = chunks.filter((c) => c.status === 'preview').length;
-  const errorCount = chunks.filter((c) => c.status === 'error').length;
-  const progressPct = total > 0 ? Math.round((completedCount / total) * 100) : 0;
-  const compositeQuality = calculateCompositeQuality(chunks);
-  const compositeLabel = compositeQuality ? t(qualityLabelKey(compositeQuality)) : null;
-  const compositeTone = qualityTone(compositeQuality);
-  const totalTokens = usageSummary.overall.totalInput + usageSummary.overall.totalOutput;
-
-  if (chunks.length === 0) {
-    return (
-      <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="flex flex-col items-center justify-center gap-3 px-6 py-12 text-center">
-        <BarChart2 size={28} className="text-editorial-border" />
-        <p className="text-sm font-medium text-editorial-muted">{t('document.indexEmptyTitle')}</p>
-        <p className="text-xs leading-relaxed text-editorial-muted/70">{t('document.indexEmptyBody')}</p>
-      </div>
-    );
-  }
-
-  return (
-    <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="space-y-3 px-5 py-5">
-      <section className="rounded-[20px] border border-editorial-border bg-editorial-bg px-4 py-3">
-        <div className="mb-3 flex items-center gap-1.5 text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">
-          <FileText size={11} className="text-editorial-accent shrink-0" /> {t('document.infoLabel')}
-        </div>
-        <dl className="space-y-2">
-          <StatRow label={t('document.infoSourceWords')} value={sourceWords.toLocaleString()} />
-          <StatRow label={t('document.infoTranslatedWords')} value={`${translatedWords.toLocaleString()} (${coverageRatio}%)`} />
-          <StatRow label={t('document.infoChunks')} value={`${completedCount} / ${total}`} />
-        </dl>
-      </section>
-
-      <section className="rounded-[20px] border border-editorial-border bg-editorial-bg px-4 py-3">
-        <div className="mb-3 flex items-center gap-1.5 text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">
-          <BarChart2 size={11} className="text-editorial-accent shrink-0" /> {t('pipeline.chunkStatus.completed')}
-        </div>
-        <div className="mb-3 h-1.5 w-full overflow-hidden rounded-full bg-editorial-border/40">
-          <div className="h-full rounded-full bg-editorial-success transition-all" style={{ width: `${progressPct}%` }} />
-        </div>
-        <div className="mb-2 font-display text-lg italic text-editorial-ink">{progressPct}%</div>
-        <div className="flex flex-wrap gap-3">
-          {idleCount > 0 && <div className="flex items-center gap-1.5 text-[10px] text-editorial-muted"><Circle size={10} className="text-editorial-muted/50" /><span className="font-bold">{idleCount}</span> {t('pipeline.chunkStatus.ready')}</div>}
-          {processingCount > 0 && <div className="flex items-center gap-1.5 text-[10px] text-editorial-warning"><Loader2 size={10} className="animate-spin" /><span className="font-bold">{processingCount}</span> {t('pipeline.chunkStatus.processing')}</div>}
-          {previewCount > 0 && <div className="flex items-center gap-1.5 text-[10px] text-editorial-muted"><FlaskConical size={10} /><span className="font-bold">{previewCount}</span> {t('pipeline.chunkStatus.preview')}</div>}
-          {completedCount > 0 && <div className="flex items-center gap-1.5 text-[10px] text-editorial-success"><CheckCircle2 size={10} /><span className="font-bold">{completedCount}</span> {t('pipeline.chunkStatus.completed')}</div>}
-          {errorCount > 0 && <div className="flex items-center gap-1.5 text-[10px] text-editorial-accent"><AlertCircle size={10} /><span className="font-bold">{errorCount}</span> {t('pipeline.chunkStatus.error')}</div>}
-        </div>
-      </section>
-
-      <section className="rounded-[20px] border border-editorial-border bg-editorial-bg px-4 py-3">
-        <div className="mb-3 flex items-center gap-1.5 text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">
-          <Gauge size={11} className="text-editorial-accent shrink-0" /> {t('document.infoQuality')}
-        </div>
-        {compositeLabel
-          ? <div className={`font-display text-lg italic ${QUALITY_TONE_COLOR[compositeTone]}`}>{compositeLabel}</div>
-          : <div className="font-display text-lg italic text-editorial-muted/40">—</div>}
-      </section>
-
-      <section className="rounded-[20px] border border-editorial-border bg-editorial-bg px-4 py-3">
-        <div className="mb-3 flex items-center gap-1.5 text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">
-          <Cpu size={11} className="text-editorial-accent shrink-0" /> {t('header.tokenCount')}
-        </div>
-        <dl className="space-y-2">
-          <StatRow
-            label={t('header.tokenCount')}
-            value={totalTokens > 0 ? totalTokens.toLocaleString() : '—'}
-          />
-          {totalTokens > 0 && (
-            <div className="flex items-baseline gap-1 pl-3">
-              <dt className="text-[10px] font-bold uppercase tracking-[0.2em] text-editorial-muted/60">in</dt>
-              <dd className="font-display text-sm italic text-editorial-muted">{usageSummary.overall.totalInput.toLocaleString()}</dd>
-              <dt className="ml-2 text-[10px] font-bold uppercase tracking-[0.2em] text-editorial-muted/60">out</dt>
-              <dd className="font-display text-sm italic text-editorial-muted">{usageSummary.overall.totalOutput.toLocaleString()}</dd>
-            </div>
-          )}
-          {(usageSummary.overall.totalCached > 0 || usageSummary.overall.totalCacheMiss > 0) && (
-            <>
-              <StatRow label={t('header.cachedInput')} value={usageSummary.overall.totalCached.toLocaleString()} />
-              <StatRow label={t('header.cacheHitRate')} value={formatCacheHitRate(usageSummary.overall.cacheHitRate)} />
-              {usageSummary.overall.totalCacheMiss > 0 && (
-                <StatRow label={t('header.cacheMissInput')} value={usageSummary.overall.totalCacheMiss.toLocaleString()} />
-              )}
-            </>
-          )}
-          <StatRow label={t('header.estimatedCost')} value={formatUsd(usageSummary.overall.totalUsd)} />
-          <StatRow label={t('log.totalDuration')} value={usageSummary.overall.totalDurationMs > 0 ? formatDurationMs(usageSummary.overall.totalDurationMs) : '—'} />
-          <StatRow label={t('document.summaryTranslationRuns')} value={usageSummary.translationRuns.toLocaleString()} />
-          <StatRow label={t('document.summaryAuditRuns')} value={usageSummary.auditRuns.toLocaleString()} />
-        </dl>
-        {usageSummary.scopeBreakdown.length > 0 && (
-          <div className="mt-3 space-y-1">
-            {usageSummary.scopeBreakdown.map((entry, i) => (
-              <ScopeBreakdownCard key={`${entry.scope}-${i}`} entry={entry} />
-            ))}
-          </div>
-        )}
-      </section>
-    </div>
-  );
-}
-
-function StatRow({ label, value }: { label: string; value: string }) {
-  return (
-    <div className="flex items-baseline gap-1">
-      <dt className="text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">{label}</dt>
-      <dd className="font-display text-sm italic text-editorial-ink">{value}</dd>
-    </div>
-  );
-}
-
-const TOP_SCOPE_I18N_KEYS = new Set(['log.scopeAudit', 'log.scopeCoherence']);
-
-function ScopeBreakdownCard({ entry }: { entry: ScopeBreakdownEntry }) {
-  const { t } = useTranslation();
-  const { labelKey, model, stats } = entry;
-  const label = TOP_SCOPE_I18N_KEYS.has(labelKey) ? t(labelKey) : labelKey;
-  const totalTok = stats.totalInput + stats.totalOutput;
-  return (
-    <div className="rounded-xl border border-editorial-border/70 bg-editorial-textbox/40 px-3 py-2.5">
-      <div className="flex items-baseline justify-between gap-2">
-        <span className="text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted shrink-0">{label}</span>
-        <span className="font-display text-sm italic text-editorial-ink shrink-0">{totalTok.toLocaleString()} tok</span>
-      </div>
-      {model && (
-        <div className="mt-0.5 truncate font-mono text-xs text-editorial-muted/70">{model}</div>
-      )}
-      <div className="mt-2 flex gap-5">
-        <div>
-          <dt className="text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">{t('header.cacheHitRate')}</dt>
-          <dd className="font-display text-sm italic text-editorial-ink">{formatCacheHitRate(stats.cacheHitRate)}</dd>
-        </div>
-        <div>
-          <dt className="text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">{t('header.estimatedCost')}</dt>
-          <dd className="font-display text-sm italic text-editorial-ink">{formatUsd(stats.totalUsd)}</dd>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-interface ChunkSummaryTabProps {
-  panelId: string;
-  labelledBy: string;
-  currentChunk: TranslationChunk | null;
-}
-
-function ChunkSummaryTab({ panelId, labelledBy, currentChunk }: ChunkSummaryTabProps) {
-  const { t, i18n } = useTranslation();
-  const pricingOverrides = usePricingStore((s) => s.overrides);
-  const logEntries = useOperationLogStore((s) => s.entries);
-  const chunkSummary = useMemo(
-    () => (currentChunk ? summarizeChunkUsage(logEntries, currentChunk.id, pricingOverrides) : null),
-    [currentChunk, logEntries, pricingOverrides],
-  );
-
-  if (!currentChunk || !chunkSummary) {
-    return (
-      <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="px-6 py-8 text-sm text-editorial-muted">
-        {t('document.insightsSummaryEmpty')}
-      </div>
-    );
-  }
-
-  const totalTokens = chunkSummary.total.totalInput + chunkSummary.total.totalOutput;
-
-  return (
-    <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="space-y-3 px-5 py-5">
-      <section className="rounded-[20px] border border-editorial-border bg-editorial-bg px-4 py-3">
-        <div className="mb-3 flex items-center gap-1.5 text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">
-          <BarChart2 size={11} className="text-editorial-accent shrink-0" /> {t('document.chunkSummaryTotals')}
-        </div>
-        <dl className="space-y-2">
-          <StatRow label={t('header.tokenCount')} value={totalTokens > 0 ? totalTokens.toLocaleString() : '—'} />
-          <StatRow label={t('header.cachedInput')} value={chunkSummary.total.totalCached.toLocaleString()} />
-          <StatRow label={t('header.cacheHitRate')} value={formatCacheHitRate(chunkSummary.total.cacheHitRate)} />
-          <StatRow label={t('header.estimatedCost')} value={formatUsd(chunkSummary.total.totalUsd)} />
-          <StatRow label={t('log.totalDuration')} value={chunkSummary.total.totalDurationMs > 0 ? formatDurationMs(chunkSummary.total.totalDurationMs) : '—'} />
-          <StatRow label={t('document.summaryTranslationRuns')} value={chunkSummary.translationRuns.toLocaleString()} />
-          <StatRow label={t('document.summaryAuditRuns')} value={chunkSummary.auditRuns.toLocaleString()} />
-        </dl>
-      </section>
-
-      <RunSummaryCard
-        title={t('document.chunkSummaryLastTranslation')}
-        emptyLabel={t('document.chunkSummaryNoTranslation')}
-        run={chunkSummary.lastTranslationRun}
-        locale={i18n.language}
-      />
-
-      <RunSummaryCard
-        title={t('document.chunkSummaryLastAudit')}
-        emptyLabel={t('document.chunkSummaryNoAudit')}
-        run={chunkSummary.lastAuditRun}
-        locale={i18n.language}
-      />
-    </div>
-  );
-}
-
-function RunSummaryCard({
-  title,
-  emptyLabel,
-  run,
-  locale,
-}: {
-  title: string;
-  emptyLabel: string;
-  run: OperationLogRunSummary | null;
-  locale: string;
-}) {
-  const { t } = useTranslation();
-
-  return (
-    <section className="rounded-[20px] border border-editorial-border bg-editorial-bg px-4 py-3">
-      <div className="mb-3 flex items-center gap-1.5 text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">
-        <Cpu size={11} className="text-editorial-accent shrink-0" /> {title}
-      </div>
-      {!run ? (
-        <p className="text-[11px] leading-relaxed text-editorial-muted/70">{emptyLabel}</p>
-      ) : (
-        <dl className="space-y-2">
-          <StatRow label={t('document.summaryWhen')} value={new Date(run.at).toLocaleString(locale)} />
-          {run.stageName ? <StatRow label={t('document.summaryStage')} value={run.stageName} /> : null}
-          <StatRow label={t('document.summaryModel')} value={[run.provider, run.model].filter(Boolean).join(' / ') || '—'} />
-          <StatRow label={t('document.summaryTokens')} value={(run.stats.totalInput + run.stats.totalOutput).toLocaleString()} />
-          <StatRow label={t('document.summaryCached')} value={run.stats.totalCached.toLocaleString()} />
-          <StatRow label={t('header.cacheHitRate')} value={formatCacheHitRate(run.stats.cacheHitRate)} />
-          <StatRow label={t('document.summaryMiss')} value={run.stats.totalCacheMiss.toLocaleString()} />
-          <StatRow label={t('log.totalDuration')} value={run.stats.totalDurationMs > 0 ? formatDurationMs(run.stats.totalDurationMs) : '—'} />
-        </dl>
-      )}
-    </section>
-  );
-}
-
-// ── Coherence Tab ──────────────────────────────────────────────────────────
-
-interface CoherenceTabProps {
-  panelId: string;
-  labelledBy: string;
-  currentChunk: TranslationChunk | null;
-  isProcessing: boolean;
-  allChunksTranslated: boolean;
-  allChunksLocked: boolean;
-  unlockedChunksCount: number;
-  onSelectChunk: (id: string) => void;
-  onFocusIssue: (chunkId: string, query?: string | null) => void;
-  onRunCoherenceAudit: () => void;
-}
-
-function CoherenceTab({ panelId, labelledBy, currentChunk, isProcessing, allChunksTranslated, allChunksLocked, unlockedChunksCount, onSelectChunk, onFocusIssue, onRunCoherenceAudit }: CoherenceTabProps) {
-  const { t } = useTranslation();
-  const coherence = currentChunk?.coherenceResult;
-  const coherenceDisabled = isProcessing || !allChunksTranslated;
-  const coherenceTitle = coherenceDisabled && !isProcessing
-    ? t('coherence.translationsRequired')
-    : coherence?.status === 'completed' || coherence?.status === 'error'
-      ? t('coherence.rerun')
-      : t('coherence.runAudit');
-
-  return (
-    <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="px-5 py-5">
-      <section className="rounded-[20px] border border-editorial-border bg-editorial-bg p-5">
-        <div className="flex items-center justify-between gap-3">
-          <div className="flex items-center gap-2 text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">
-            <Link2 size={11} className="text-editorial-accent shrink-0" /> {t('coherence.title')}
-          </div>
-          <IconButton
-            size="md"
-            onClick={onRunCoherenceAudit}
-            disabled={coherenceDisabled}
-            title={coherenceTitle}
-            ariaLabel={t('coherence.runAudit')}
-            tooltipSide="left"
-          >
-            {coherence?.status === 'processing' ? <Loader2 size={14} className="animate-spin" /> : <ScanLine size={14} />}
-          </IconButton>
-        </div>
-
-        {allChunksTranslated && !allChunksLocked && (
-          <div className="mt-3 flex items-start gap-2 rounded-2xl border border-editorial-warning/30 bg-editorial-warning/10 p-3 text-sm text-editorial-warning">
-            <AlertTriangle size={14} className="mt-0.5 shrink-0" />
-            <span>{t('coherence.unlockedWarning', { count: unlockedChunksCount })}</span>
-          </div>
-        )}
-
-        {!coherence || coherence.status === 'idle' ? (
-          <p className="mt-3 text-[11px] text-editorial-muted/70 leading-relaxed">
-            {!allChunksTranslated ? t('coherence.translationsRequired') : t('coherence.idle')}
-          </p>
-        ) : coherence.status === 'processing' ? (
-          <div className="mt-3 flex items-center gap-2 text-sm text-editorial-muted">
-            <Loader2 size={13} className="animate-spin shrink-0" /> {t('coherence.running')}
-          </div>
-        ) : coherence.status === 'error' ? (
-          <div className="mt-3 rounded-2xl border border-editorial-accent/30 bg-editorial-textbox/40 p-3 text-sm text-editorial-accent">
-            {coherence.error || t('errors.coherenceFailed')}
-          </div>
-        ) : coherence.issues.length === 0 ? (
-          <div className="mt-3 flex items-center gap-2 text-sm text-editorial-success">
-            <CheckCircle2 size={14} /> {t('coherence.noIssues')}
-          </div>
-        ) : currentChunk ? (
-          <IssueList issues={coherence.issues} chunkId={currentChunk.id} onSelectChunk={onSelectChunk} onFocusIssue={onFocusIssue} />
-        ) : null}
-      </section>
-    </div>
-  );
-}
-
-// ── Prompt viewer (inline) ─────────────────────────────────────────────────
-
-
-// ── Audit Tab ──────────────────────────────────────────────────────────────
-
-interface AuditTabProps {
-  panelId: string;
-  labelledBy: string;
-  currentChunk: TranslationChunk | null;
-  isProcessing: boolean;
-  onReauditChunk: (chunkId: string) => void;
-  onSelectChunk: (id: string) => void;
-  onFocusIssue: (chunkId: string, query?: string | null) => void;
-}
-
-function AuditTab({ panelId, labelledBy, currentChunk, isProcessing, onReauditChunk, onSelectChunk, onFocusIssue }: AuditTabProps) {
-  const { t } = useTranslation();
-
-  if (!currentChunk) {
-    return (
-      <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="px-6 py-8 text-sm text-editorial-muted">
-        {t('document.insightsAuditEmpty')}
-      </div>
-    );
-  }
-
-  const tone = qualityTone(currentChunk.judgeResult.rating);
-  const qualityLabel = currentChunk.judgeResult.status === 'completed'
-    ? t(qualityLabelKey(currentChunk.judgeResult.rating))
-    : t('audit.ratingNone');
-
-  return (
-    <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="px-5 py-5">
-      <section className="rounded-[20px] border border-editorial-border bg-editorial-bg p-5">
-        <div className="flex items-start justify-between gap-3">
-          <div>
-            <div className="text-[10px] font-sans uppercase tracking-[0.35em] text-editorial-muted">{t('audit.title')}</div>
-            <div className={`mt-2 font-display text-xl italic ${QUALITY_TONE_COLOR[tone]}`}>{qualityLabel}</div>
-          </div>
-          <IconButton
-            size="md"
-            onClick={() => onReauditChunk(currentChunk.id)}
-            disabled={isProcessing || !currentChunk.currentDraft}
-            title={t('pipeline.reauditChunk')}
-            tooltipSide="left"
-          >
-            <RefreshCcw size={14} />
-          </IconButton>
-        </div>
-
-        {currentChunk.judgeResult.status === 'error' && (
-          <div className="mt-4 rounded-2xl border border-editorial-accent/30 bg-editorial-textbox/40 p-4 text-sm text-editorial-accent">
-            {currentChunk.judgeResult.error || t('audit.auditFailed')}
-          </div>
-        )}
-        {currentChunk.judgeResult.status !== 'error' && currentChunk.judgeResult.status !== 'completed' && (
-          <div className="mt-4 rounded-2xl border border-editorial-border bg-editorial-bg/60 p-4 text-sm text-editorial-muted">
-            {t('document.insightsAuditEmpty')}
-          </div>
-        )}
-        {currentChunk.judgeResult.status === 'completed' && currentChunk.judgeResult.issues.length === 0 && (
-          <div className="mt-4 flex items-center gap-2 text-sm text-editorial-success">
-            <CheckCircle2 size={14} /> {t('document.insightsAuditNoIssues')}
-          </div>
-        )}
-        {currentChunk.judgeResult.issues.length > 0 && (
-          <IssueList issues={currentChunk.judgeResult.issues} chunkId={currentChunk.id} onSelectChunk={onSelectChunk} onFocusIssue={onFocusIssue} />
-        )}
-      </section>
-    </div>
-  );
-}
-
-// ── Notes Tab ──────────────────────────────────────────────────────────────
-
-interface NotesTabProps {
-  panelId: string;
-  labelledBy: string;
-  currentChunk: TranslationChunk | null;
-}
-
-function NotesTab({ panelId, labelledBy, currentChunk }: NotesTabProps) {
-  const { t } = useTranslation();
-  const footnotes = currentChunk?.footnotes ?? [];
-
-  if (!currentChunk || footnotes.length === 0) {
-    return (
-      <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="flex flex-col items-center justify-center gap-3 px-6 py-12 text-center">
-        <NotebookText size={28} className="text-editorial-border" />
-        <p className="text-sm font-medium text-editorial-muted">{t('document.insightsNotesEmpty')}</p>
-      </div>
-    );
-  }
-
-  return (
-    <div id={panelId} role="tabpanel" aria-labelledby={labelledBy} className="space-y-2 px-5 py-5">
-      {footnotes.map((note) => (
-        <article key={note.id} className="rounded-2xl border border-editorial-border bg-editorial-bg px-4 py-3">
-          <div className="mb-1.5 font-display text-sm italic text-editorial-accent">{note.marker}</div>
-          <p className="text-[12px] leading-relaxed text-editorial-ink">{note.text}</p>
-        </article>
-      ))}
-    </div>
-  );
-}
-
-// ── IssueList ──────────────────────────────────────────────────────────────
-
-interface IssueListProps {
-  issues: TranslationChunk['judgeResult']['issues'];
-  chunkId: string;
-  onSelectChunk: (id: string) => void;
-  onFocusIssue: (chunkId: string, query?: string | null) => void;
-}
-
-function IssueList({ issues, chunkId, onSelectChunk, onFocusIssue }: IssueListProps) {
-  const { t } = useTranslation();
-  const focusedIssueQuery = useUiStore((s) => s.focusedIssueQuery);
-  const clearFocusedIssue = useUiStore((s) => s.clearFocusedIssue);
-  return (
-    <div className="mt-4 space-y-3">
-      {issues.map((issue, index) => (
-        <article key={`${issue.type}-${index}`} className="rounded-2xl border border-editorial-border bg-editorial-bg/80 p-4 shadow-sm">
-          <div className="mb-3 flex items-center justify-between gap-3">
-            <div className="flex items-center gap-2">
-              <span className={`rounded-full p-1 ${issue.severity === 'high' ? 'bg-editorial-accent text-white' : issue.severity === 'medium' ? 'bg-editorial-warning/80 text-white' : 'bg-editorial-border text-editorial-muted'}`}>
-                {issue.type === 'fluency' ? <MessageCircle size={11} /> :
-                 issue.type === 'accuracy' ? <AlertTriangle size={11} /> :
-                 issue.type === 'grammar' ? <AlertCircle size={11} /> :
-                 issue.type === 'consistency' ? <Link2 size={11} /> :
-                 <BookOpen size={11} />}
-              </span>
-              <span className="text-[10px] font-bold uppercase tracking-[0.2em] text-editorial-ink">{issue.type}</span>
-            </div>
-            {issue.phrase && (() => {
-              const isActive = focusedIssueQuery === issue.phrase;
-              return (
-                <IconButton
-                  size="sm"
-                  tone={isActive ? 'accent' : 'default'}
-                  onClick={() => {
-                    if (isActive) {
-                      clearFocusedIssue();
-                    } else {
-                      onSelectChunk(chunkId);
-                      onFocusIssue(chunkId, issue.phrase);
-                    }
-                  }}
-                  title={t('audit.locateInTextTooltip')}
-                  ariaPressed={isActive}
-                  tooltipSide="left"
-                >
-                  <Crosshair size={13} />
-                </IconButton>
-              );
-            })()}
-          </div>
-          <p className="text-sm leading-relaxed text-editorial-ink">{issue.description}</p>
-          {issue.suggestedFix && (
-            <div className="mt-3 rounded-xl border border-editorial-border/70 bg-editorial-bg px-3 py-2 text-sm leading-relaxed text-editorial-muted">
-              <span className="text-[10px] font-bold uppercase tracking-[0.2em] text-editorial-accent">{t('audit.fix')}</span>: {issue.suggestedFix}
-            </div>
-          )}
-        </article>
-      ))}
-    </div>
-  );
-}
-
-// ── Helpers ────────────────────────────────────────────────────────────────
-
-function countWords(text: string): number {
-  return text.trim().split(/\s+/).filter(Boolean).length;
-}
-
-
-// ── Glossary Tab ────────────────────────────────────────────────────────────
-
-interface GlossaryTabProps {
-  panelId: string;
-  labelledBy: string;
-  glossary: Array<{ id?: string; term: string; translation: string; notes?: string }>;
-}
-
-function GlossaryTab({ panelId, labelledBy, glossary }: GlossaryTabProps) {
-  const { t } = useTranslation();
-  return (
-    <div
-      id={panelId}
-      role="tabpanel"
-      aria-labelledby={labelledBy}
-      className="flex flex-1 flex-col"
-    >
-      <div className="border-b border-editorial-border px-5 py-3">
-        <span className="text-xs font-mono text-editorial-muted">{glossary.length} {t('document.insightsTabGlossary').toLowerCase()}</span>
-      </div>
-      <div className="flex-1 overflow-y-auto custom-scrollbar px-4 py-2">
-        <table className="w-full text-sm">
-          <tbody>
-            {glossary.map((entry, i) => (
-              <tr key={entry.id ?? i} className="border-b border-editorial-border/40 last:border-0">
-                <td className="py-2 pr-3 font-medium text-editorial-ink">{entry.term}</td>
-                <td className="py-2 text-editorial-muted/60">→</td>
-                <td className="py-2 pl-3 text-editorial-ink">{entry.translation}</td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
     </div>
   );
 }
